Move betty.plugins.copyright_notice to betty.copyright_notices
