+++ betty/copyright_notices/__init__.py
