+++ betty/copyright_notices/project_author.py
