+++ betty/copyright_notices/public_domain.py
