+++ betty/copyright_notices/wikipedia_contributors.py
