+++ betty/loaders/demo/__init__.py
@@ -6,12 +6,12 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Self, final, override
 
+from betty.copyright_notices.streetmix import Streetmix
 from betty.factory import Manufacturable
 from betty.license import LicenseDefinition
 from betty.licenses.spdx import spdx_license_id_to_license_id
 from betty.load import Loader, LoaderDefinition
 from betty.loaders.demo.jobs import LoadAncestry
-from betty.plugins.copyright_notice.streetmix import Streetmix
 from betty.project import Project
 
 if TYPE_CHECKING:
