+++ betty/plugins/extension/wiki/__init__.py
@@ -5,9 +5,9 @@ from __future__ import annotations
 from asyncio import gather
 from typing import Self, final, override
 
+from betty.copyright_notices.wikipedia_contributors import WikipediaContributors
 from betty.extension import Extension, ExtensionDefinition
 from betty.factory import Manufacturable
-from betty.plugins.copyright_notice.wikipedia_contributors import WikipediaContributors
 from betty.project import Project
 from betty.service import ServiceProvider
 from betty.service.simple import service
