+++ betty/tests/copyright_notices/test_project_author.py
@@ -3,7 +3,7 @@ from typing import override
 import pytest
 
 from betty.copyright_notice import CopyrightNotice
-from betty.plugins.copyright_notice.project_author import ProjectAuthor
+from betty.copyright_notices.project_author import ProjectAuthor
 from betty.test_utils.copyright_notice import CopyrightNoticeTestBase
