+++ betty/tests/copyright_notices/test_streetmix.py
@@ -3,7 +3,7 @@ from typing import override
 import pytest
 
 from betty.copyright_notice import CopyrightNotice
-from betty.plugins.copyright_notice.streetmix import Streetmix
+from betty.copyright_notices.streetmix import Streetmix
 from betty.test_utils.copyright_notice import CopyrightNoticeTestBase
