+++ pyproject.toml
@@ -128,11 +128,11 @@ betty = 'betty.console:main_standalone'
 'update-translations' = 'betty.plugins.command.update_translations:UpdateTranslations'
 
 [project.entry-points.'betty.copyright_notice']
-'project-author' = 'betty.plugins.copyright_notice.project_author:ProjectAuthor'
+'project-author' = 'betty.copyright_notices.project_author:ProjectAuthor'
 'project-configuration' = 'betty.project._plugin_discovery:_discover_copyright_notices'
-'public-domain' = 'betty.plugins.copyright_notice.public_domain:PublicDomain'
-'streetmix' = 'betty.plugins.copyright_notice.streetmix:Streetmix'
-'wikipedia-contributors' = 'betty.plugins.copyright_notice.wikipedia_contributors:WikipediaContributors'
+'public-domain' = 'betty.copyright_notices.public_domain:PublicDomain'
+'streetmix' = 'betty.copyright_notices.streetmix:Streetmix'
+'wikipedia-contributors' = 'betty.copyright_notices.wikipedia_contributors:WikipediaContributors'
 
 [project.entry-points.'betty.css_resource']
 'webpack' = 'betty.plugins.css_resource.webpack:WEBPACK'
